Layer関係のclassを別ファイルに切り出す
