+++ front/composables/Layers.ts
@@ -1,0 +1,4 @@
+import { ColorImage } from "./ColorImage"
+
+export type Layer =
+  | ColorImage
